[Refactor Fee][Nam] Update create request of Subscription Plan

+++ Vigig.Service/Models/Request/SubscriptionPlan/CreateSubscriptionPlanRequest.cs
@@ -10,7 +10,7 @@ public class CreateSubscriptionPlanRequest
 
     [Required]
     [MinValue(1)]
-    public int? DurationType { get; set; }
+    public double? DurationType { get; set; }
 
     [Required]
     [MinValue(0)]
